pyproject.toml
[project]
name = "urlshorter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.121.3",
    "httpx>=0.28.1",
    "mongo>=0.2.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "uvloop>=0.21.0; sys_platform != 'win32'", # для скорости работы fastapi
]
